custom_components/whisper-asr/requirements.txt
openai>=1.0.0
redis>=4.5.0
pyaudio>=0.2.13
numpy>=1.20.0
python-dotenv>=1.0.0
